Changed how Students are built while changing phone back to int
This was the start of the Builder Design Pattern, but was simplified and became an implementation of the Factory Design Pattern.

It almost looks useless in this case, but it actually makes for a cleaner 'constructor' code.

Phone went back to int, according to unit test. No reason to go for long.