Improvements on Student and test case.
Most comment changes were about a better english.

generateNumber was renamed to seed because generate is a verb, which is reserved for methods.

Unchecked exceptions are preferred (IllegalArgumentException or RuntimeException instead of Exception), because of maintenance issues. Checks moved to top to render else { } unnecessary which makes for cleaner code.

Id concatenation was extracted to a default constructor for testing purposes. Note a test was written to test an edge case of a full seed, but it's not working. We need to rewrite it.