Renamed MealMenu to just Menu
It's simpler. The reason it was MealMenu before, was to not confuse with a possible graphical interface menu object, but since it's on a separate package, there's no conflict.

The motivation behind the rename was to add another object MealMenuIdentifier to uniquely identify a Menu object. The name was getting too long.