Draft implementation of DishMapper
Seems to be working.

This is a simple mapper, used for the list of dishes (i.e., meat, fish, soups…). Since they're simple strings, this mapper doesn't map a domain object. It maps String objects.

No documentation yet, until it's refactored and stable. This avoids wasting time documenting stuff that will be deleted or outdated soon.