Moved loaded instance handling and default constructor parameter from each mapper to a new MapperRegistry class.
The MapperRegistry class is to be used whenever a data mapper is needed (e.g., MapperRegistry.menu() for a MenuMapper instance). It will only be created the first time it's needed (lazy loading).

This removes unwanted duplication and does not violate SRP (Single Responsibility Principle) in mappers.

Methods to more easily change are located at the top of the file for easy editing, such as the default parent data path for files or the location of the accounts data file.