Complete rewrite of Application. Added Student and Administrator management workflows, while fixing bugs.
Application is now up to date with Student and Administrator management needs. There are gateway methods to create a default Administrator account, create a new student, close an account and authenticate users.

StudentMapper and related objects add a few bugs that were fixed.

Added coupling in Student for Account. It's not mandatory to create a Student with an Account, but if one is not present, it will be created, loaded and updated automatically. It can also be managed independently, without a Student object.

Note that id fields need to be an object (Integer), because if we save a null to an int, it will be saved as a 0 (zero), and it makes checking if an id is present (for newly created objects) difficult.

Build methods were removed in favor of a regular constructor, in anticipation of validation being done in separate classes in the near future.