Complete rewrite of Menu and related objects.
It was a requirement to store in an xml file the menu for a day, including meals for lunch and dinner. The previous version of Menu only supported a menu for a time of day.

This is a simpler solution because it's in line with the file, it's more dynamic because it doesn't assume enum values. Besides, the previous Menu implementation with Menu.Id was a bit weird, compared to the rest.

A Day class was added to make it easier to use dates disregarding time.